rtc: send absolute offset from UTC to QEMU.
For one, update the RTC when guest returns from PM state, that part
should be important to keep RTC time consistent within guest. The second
part about adjustment does not seem to bring much except sending QEMU
the TIMEOFFSET IOREQ before updating the domain wallclock. There is a
patch in QEMU to store diff for this guest in Xenstore to have it
persistent accross reboots. Xen will be notified of this at QEMU's start
through xc_domain_set_time_offset().

It is not clear exactly where this patch originated. It has a component
that goes along with it in QEMU (maintain-time-offset.patch). The short
description is all the original information that can be found. The above
description is the analysis of the patch during the 4.6.1 uprev about it
likely purpose.

Signed-off-by: unknown
[Moved in OpenXT layer]
Signed-off-by: Philip Tricca <flihp@twobit.us>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Documented]
Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>