rtc: uptade current time processing request.
Fix problem where time jumps forward in windows guest. Ensure that the
current_tm field is updated before processing a request to modify time
as this is used to calculate the delta between the current guest time
and the new time.

It is unclear in what conditions that was required specificaly.

Signed-off-by: unknown
[Moved in OpenXT layer]
Signed-off-by: Philip Tricca <flihp@twobit.us>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Documented]
Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>