acpi: PM features for guest battery, ac, lid.
This provides enhanced PM capabilities for batteries, AC adapters and
laptop lids. It also supports power and sleep button even injection. It
depends on the acpi-pm-feature.patch in the QEMU patch queue. The two
parts work in conjunction to surface various PM related devices in the
PM SSDT. There is a partial version of this in upstream Xen that was
removed recently.

Signed-off-By: Kamala Narasimhan <kamala.narasimhan@citrix.com>
[Refactored]
Signed-off-By: Nicolas DI PRIMA <nicolas.diprima@citrix.com>
[Fixed, refactored and rebased on Xen 4.3.4]
Signed-off-By: Ross Philipson <philipsonr@ainfosec.com>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Documented]
Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>