acpi: work-around Dell 980 broken firmware.
The PM1a_CNT block on the Dell looks like:
[0ACh 0172  12]           PM1A Control Block : <Generic Address Structure>
[0ACh 0172   1]                     Space ID : 01 (SystemIO)
[0ADh 0173   1]                    Bit Width : 10
[0AEh 0174   1]                   Bit Offset : 00
[0AFh 0175   1]         Encoded Access Width : 01 (Byte Access:8)
[0B0h 0176   8]                      Address : 0000000000000804

The access width is 1 (byte).  However, doing a single byte out causes
the hang.  So Dell's ACPI table is incorrect.

The reason this works for Xen w/o tboot is that Xen ignores the access
width and always does a 16 bit outw--which works.  But according to one
of our ACPI maintainers, the code really should respect the access width
value.

If I run Linux w/ tboot it appears that Linux is changing the access
width value to 0 and that defaults to 16 bits--which works.

Signed-off-by: unknown
[Moved in OpenXT layer]
Signed-off-by: Philip Tricca <flihp@twobit.us>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>