cpuid: fake CPUID returned value.
1- XciVMMXciVMM signature was used by XenClient old custom Linux PV
drivers to get them to load instead of standard pvops ones. This is kept
because it is very useful for debugging. By enabling this CPUID one can
force a pvops guest to not load its standard PV drivers.

2- Exposes VCPUs as multiple cores in a smaller number of sockets, by
adjusting the CPUID responses appropriately. This is to make Windows
think that VCPUs were cores not CPUs so OpenXT could overcome CPU count
limits and licensing issues in Windows. The addition op
XEN_DOMCTL_setcorespersocket is the DOMCLT op that configures Xen to
deal with the differences between cores and CPUs.

Signed-off-by: unknown
[Moved in OpenXT layer]
Signed-off-by: Philip Tricca <flihp@twobit.us>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Documented, fixes]
Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>