privcmd: Add mmapcacheattr to libxc interface.
The display framebuffer requires correct memory cache attributes to be
set in order to avoid visible graphical artefacts as memory writes are
deferred until cache eviction occurs.

This provides the library method to invoke the kernel facility for
setting the cache attributes on a new mapping of guest memory.

TODO:
Upstream support for the IOCTL_PRIVCMD_MMAPCACHEATTR ioctl on the Xen
privileged command file in Linux is a hard requirement before this patch
could be considered.

Implementation of the method in this patch is currently only provided
for Linux and not the other OSes supported by Xen since OpenXT does not
require and would not exercise such code.

Signed-off-by: unknown
[Moved in OpenXT layer]
Signed-off-by: Philip Tricca <flihp@twobit.us>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>