x86/hvm: add option to disable VMCS shadowing.
This patch was created to support vSentry running as a nested VMM on
OpenXT.

TODO:
It is not entirely clear why VMCS shadowing needs to be disabled since
it is a feature to assist nested VMMs. It may not be needed any longer
to support vSentry but it will be left here as a placeholder since it
does nothing by default.