hvmloader: get SeaBIOS image from outside the tree.
This patch allows SEABIOS to be built and pulled in from outside the Xen
tree.  It also introduces a new function to layout the legacy guest
firmware for the SEABIOS ROM, the selected VGA BIOS and the Ethernet/PXE
OROM.

TODO:
Note the patch does not pop without being forced because it touches the
tools/firmare/Makefile file that is modified by the configure step.

Signed-off-by: unknown
[Moved in OpenXT layer]
Signed-off-by: Philip Tricca <flihp@twobit.us>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Documented]
Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>