blktap2: VHD marker flag support.
Added for compatibility reasons - it is tied into the changes to
dd_batmap_hdr for supporting keyhashes which are used for VHD
encryption.

TODO:
Most of the actual code can probably be removed. It was left there for now since
the status tool uses it. The important thing is to keep the char field in the
VHD struct so it is layed out correctly for the keyhash stuff. As far as can
be determined, it was used for the BED feature which is gone.

Signed-off-By: Jake Wires <Jake.Wires@citrix.com>
[Moved in OpenXT layer]
Signed-off-by: Philip Tricca <flihp@twobit.us>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Documented]
Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>