blktap2: ro tapdisk and optional pid argument.
Optional pid argument simplifies device destruction interface and the
pause/unpause interfaces. The code had to be modified to use the old list
handling because something was missing from an earlier change.

The -R option sets the TAPDISK_MESSAGE_FLAG_RDONLY allowing tapdisks to be
created and opened RO. Two new APIs to pass flags had to be added because
something was missing from an earlier change.

Signed-off-By: Jake Wires <Jake.Wires@citrix.com>
[Ported to Xen 4.6]
Signed-off-By: Ross Philipson <philipsonr@ainfosec.com>
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Sanitized]
Signed-off-by: Jed <lejosnej@ainfosec.com>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>