blktap2: VHD encryption support.
Add the current VHD encrytion mechanisms used in OpenXT as of 2016.
This version uses AES XTS encrytion on VHD blocks. The OpenSSL crypto
library is used to do the encryption/decryption. The implementation is
in drivers/block-crypto.c and in the files in drivers/crypto/*.

The previous version used the AES XTS crypto support from the Linux
kernel. All of this was left out in the patch making effort.

Signed-off-by: Philip Tricca <philip.tricca@citrix.com>
Signed-off-by: Vincent Hanquez <vincent.hanquez@eu.citrix.com>
Signed-off-by: Tomasz Wroblewski <tomasz.wroblewski@citrix.com>
Signed-off-by: Jake Wires <Jake.Wires@citrix.com>
Signed-off-by: Christian Limpach <Christian.Limpach@citrix.com>
[Moved in OpenXT layer]
Signed-off-by: Philip Tricca <flihp@twobit.us>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Documented]
Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>