x86: x86_emulate atomic operations for 4 & 2 bytes
When running x86_emulate, the code that writes the results back to guest
memory must ensure that the proper atomicity guaranteed by the
instructions being emulated is maintained. memcpy provides for atomic
quadword writes but writes the last 1-7 bytes using single byte
operations (rep movsb) so 32- and 16- bit updates are not done
atomically.

This change uses specific atomic operations if the size is 4 or 2 bytes.

Signed-off-by: unknown
[Moved in OpenXT layer]
Signed-off-by: Philip Tricca <flihp@twobit.us>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Documented]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>