x86: Allow stubomain to handle IRQ of its domain.
Support PCI passthrough with QEMU in a stubdomain without the need for
multiple ioreq servers.

Stubdomains need to be given sufficient privilege over the guest it
provides emulation so the PCI passthrough works correctly. This is done
using current->domain->target and calling irq_permit_access when the
stubdomain tries to map the IRQ/MSI.

The capability for multiple device models (the ioreq server patch) was
removed as the current implementation of it is far different from what
upstream Xen has settled on in 4.5.

OXT-118

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>
[Refactored without IOREQ]
Signed-off: Rian Quinn <quinnr@ainfosec.com>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>