vtd: workaround vtd breakage on clarkdale.
Unfortunately, quirks specific to the Clarkdale/Nehalem integrated
graphics device (IGD) do not function correctly with Xen's VT-d
implementation; causing mapping issues which break display and can
corrupt memory throughout dom0.

This workaround disables VT-d address translation for the affected IGD
/when used for display in dom0/, granting 1:1 access to the host's
memory. VT-d for passthrough and other dom0 devices is not affected.
With this patch in place, the affected systems function properly and
have working displays.

Security concern:
While it is true that, in general, disabling VT-d for a device opens
several security risks (e.g. that of a DMA attack), this patch itself is
not believed to present any significant security risk, as the affected
device is a trusted component of the processor, controlled by trusted
software (surfman) in a trusted, priveleged hardware domain (dom0).

It is worth noting that access to the card's DMA controller is gated by
SELinux policy-- and that the only software capable of interfacing with
the card (surfman) already has the permission to map arbitrary physical
pages.

Signed-off-by: Kyle J. Temkin <temkink@ainfosec.com>
[Rebased on Xen 4.6.1]
Signed-off-by: Machon Gregory <mbgrego@tycho.ncsc.mil>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>