x86/tboot: increase AP startup time for tboot.
TBOOT has the APs running in mini guests. The INIT IPI that Xen sends
causes a VMEXIT to TBOOT's handler. It puts the AP into the
Wait-for-SIPI state and VMRESUMES (of course the guest does not actually
run any longer at this point).

The Xen code was just waiting a very short period of time for all that
to happen and it just wasn't long enough. The timeout was made
considerably larger (up to the legacy INIT IPI timeout on Intel
platforms) to guarantee all APs startup.

OXT-747

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>