xenstat: disable tmem stats in the tool.
TMEM is not used in OpenXT and access to TMEM information through
hypercalls is blocked by XSM. The tool should not try to access TMEM
information.

With Kconfig, it should be possible to disable TMEM, which would be an
upstreamable alternative to this patch.

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>
[Handle compilation warning]
Christopher Clark <christopher.clark6@baesystems.com>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>