acpi: support SLIC for guest OEM install.
In order to use Windows OEM install media, the SLIC table must be passed
to a guest. In addition all the OEM table IDs must match the SLIC or
Windows will think it is invalid and the install ends up unactivated.

NOTE: The DSDT does not get updated. This was the same in the original
patch.  Not sure why that is but it is being left that way for now since
it works w/o it.

Signed-off-by: unknown
[Ported from qemu-acpi-tables.patch]
Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>
[Rebased on Xen 4.6.6]
Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>